### 2020-05-23 #2204 @jcolomb
looking for R, machine learning, statistics, (and not epidemiology) none found for Rshiny.
(PS would be cool to have a possibility to filter that list, maybe directly from the result of whedon software report ?).

nhejazi
jayrobwilliams
ati-ozgur
rcannood
