### 2020-05-23 #2204 @danielskatz
We have a bunch of ideas for potential changes in progress to make this process better, and hope to start implementing them soon
