### 2020-05-31 #2204 @danielskatz
👋 @amoeba, @brylie, @DominiqueMakowski, @mikerspencer, @aj2duncan - would one or two of you be willing to review this JOSS submission?
