### 2020-06-01 #2204 @danielskatz
Thanks to @mikerspencer and @aj2duncan  - we will go ahead with a review in a few minutes (and in a new thread after more commands to whedon).  Mike, please feel free to ask questions as needed, but I hope the fact that the process is checklist-driven will make this reasonably clear and easy.
