### 2020-06-01 #2282 @whedon
[REVIEW]: Analysing 24-hour behaviour sequence data with an Rshiny application
**Submitting author:** @jcolomb (<a href="http://orcid.org/0000-0002-3127-5520">Colomb, J.</a>)
**Repository:** <a href="https://github.com/jcolomb/HCS_analysis" target ="_blank">https://github.com/jcolomb/HCS_analysis</a>
**Version:** v0.1.3
