### 2020-06-08 #2282 @mikerspencer
@danielskatz to be clear - @aj2duncan and I work together a little. Reviewers working together isn't specifically mentioned in the conflict of interest, but I feel it's best to be transparent.
