### 2020-06-08 #2282 @danielskatz
Thanks for letting me know. This isn't 100% ideal, but given that you both likely have different backgrounds and experiences, and we're focusing on a checklist-driven process, it should be ok. Let's see how it goes.
