### 2020-06-12 jcolomb/HCS_analysis#18 @mikerspencer
packrat::restore() installation fails
Running on local machine: Ubuntu 18.04 with R 3.6.3. I get the following error during `packrat::restore()`:

    Error: package 'plogr' 0.1-1 was found, but >= 0.2.0 is required by 'dplyr'

Is this because packrat has a snapshot of R 3.6.1?

This is issue was picked up as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2282).
