### 2020-06-14 jcolomb/HCS_analysis#18 @aj2duncan
I have a similar issue, also picked up as part of openjournals/joss-reviews#2282.

```r
packrat::restore()
```
```

The command failed with output:
* installing *source* package 'nloptr' ...
** package 'nloptr' successfully unpacked and MD5 sums checked
** using staged installation

   **********************************************
   WARNING: this package has a configure script
         It probably needs manual configuration
   **********************************************

** libs
C:/Rtools/mingw_64/bin/g++ -std=gnu++11  -I"C:/PROGRA~1/R/R-36~1.3/include" -DNDEBUG          -O2 -Wall  -mtune=core2 -c dummy.cpp -o dummy.o
C:/Rtools/mingw_64/bin/gcc  -I"C:/PROGRA~1/R/R-36~1.3/include" -DNDEBUG       -I"/x64/include"   -O2 -Wall  -std=gnu99 -mtune=core2 -c init_nloptr.c -o init_nloptr.
In addition: Warning messages:
1: In packrat::restore() :
  The most recent snapshot was generated using R version 3.6.1
2: In untar(src, exdir = target, compressed = "gzip") :
  argument 'compressed' is ignored for the internal method
3: In untar(src, compressed = "gzip", list = TRUE) :
  argument 'compressed' is ignored for the internal method
```
```r
sessionInfo()
```
```
R version 3.6.3 (2020-02-29)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 18363)

Matrix products: default

locale:
[1] LC_COLLATE=English_United Kingdom.1252
[2] LC_CTYPE=English_United Kingdom.1252
[3] LC_MONETARY=English_United Kingdom.1252
[4] LC_NUMERIC=C
[5] LC_TIME=English_United Kingdom.1252

attached base packages:

[1] compiler_3.6.3   tools_3.6.3      packrat_0.4.8-56
```

As advised I then uncommented and ran the necessary lines from  *Installation_Rpackages.R*. The command
```r
devtools::install_github('jcolomb/osfr')
```
also produced an error:
```
Error: package or namespace load failed for 'osfr' in namespaceExport(ns, exports):
 undefined exports: welcome
Error: loading failed
Execution halted
*** arch - x64
Error: package or namespace load failed for 'osfr' in namespaceExport(ns, exports):
 undefined exports: welcome
Error: loading failed
Execution halted
```

All this aside, the app runs and I will keep testing to see if the app itself produces any errors.
