### 2020-06-17 jcolomb/HCS_analysis#18 @jcolomb
sorry to be slow on  this, I will try to look at it again (I did have problems I thought were cleared)
