### 2020-06-17 jcolomb/HCS_analysis#18 @jcolomb
the osf package is not necessary anymore, I will delete it from the installation file
