### 2020-06-17 32c7f1b
erase installation of the osfr package in the manual installation file
