### 2020-06-18 #2282 @aj2duncan
@danielskatz now @jcolomb has [confirmed](https://github.com/jcolomb/HCS_analysis/issues/18) a particular part of the installation isn't necessary I will push on with the review.
