### 2020-06-19 jcolomb/HCS_analysis#19 @aj2duncan
Contribution Guide
Hi @jcolomb,

As part of the review (openjournals/joss-reviews#2282) I think that you need to add some detail to your contribution guide on the repo. You have already stated that contributions are welcome but I wonder if adding a contributing.md (demo [here](https://github.com/tidyverse/dplyr/blob/master/.github/CONTRIBUTING.md)) might help others.

Thanks,

Andrew
