### 2020-06-20 jcolomb/HCS_analysis#19 @jcolomb
good point, it was not thought as an open project in the beginning, but it should now.
