### 2020-06-20 3f9965d
Create CONTRIBUTING.md
