+++ .github/CONTRIBUTING.md
@@ -1,0 +1,28 @@
+# Contributing to HCS analysis
+
+### Suggest developement and report bugs
+
+Please use the Github issues to report bugs or ask for new analysis.
+Similarly, if you would like to expand the software to new type of raw data,
+start there so that I can help you get started and other users may learn about your initiative!
+
+
+### Fixing documentation
+
+Small typos or grammatical errors in documentation may be edited directly using
+the GitHub web interface, so long as the changes are made in the _source_ file.
+
+### New analyses
+
+In order to build new analysis, I would encourage you to start by using the `analysis/master_noshiny.R` file to upload some data, you can then use the .rdata file created (this does not work with the shiny app.).
+
+You may want to have a look at `archives_notpeerreviewed⁩/testscode⁩/analyseBseq` to see how to access the raw data (that code is trying to analyse sequence of behavior).
+
+
+### Pull request process
+
+Please check github help center to learn how to create pull requests.
+
+### Code of Conduct
+
+We do not have a code of conduct, but please be nice and welcome any contribution.
