### 2020-06-22 31de3a1
Update CONTRIBUTING.md
