### 2020-06-22 jcolomb/HCS_analysis#19 @jcolomb
done: https://github.com/jcolomb/HCS_analysis/commit/31de3a1ba4c362ca0e4037691567e66d4223cc50.

please comment if you find it not sufficient
