+++ analysis/Installation_Rpackages.R
@@ -1,18 +1,18 @@
 #installation:
-
 packrat::restore()
 
 
 ## alternatively you may install all packages and hope it works without packrat:
 #
+
 # install.packages(c(
 #   "randomForest",
 #   "ica",
 #   "e1071", #svm
-#   "Hmisc",   #binomial confidence 
-#   
-#   "rstatix"
-#   "coin"
+#   "Hmisc",   #binomial confidence
+# 
+#   "rstatix",
+#   "coin",
 #   #normal libraries:
 #   "tidyverse",
 #   "stringr",
@@ -25,8 +25,10 @@ packrat::restore()
 #   "glmpath",
 #   "pander",
 #   "DT",
-#   "beepr"))
-# 
+#   "beepr",
+#   "osfr"))
+
+
 
 ### old dependencies, no need to get it anymore
 ### devtools::install_github('jcolomb/osfr')
