### 2020-06-23 jcolomb/HCS_analysis#18 @jcolomb
still working on packrat...
osfr package actually necessary for metadata upload, I made some changes in the master branch to cope with the CRAN version of osfr.
