### 2020-07-01 #2282 @aj2duncan
@jcolomb quick query - In the shiny app and article you make reference to the 18 Berlin categories. Should this have a reference?
