### 2020-07-01 jcolomb/HCS_analysis#20 @aj2duncan
Clarification of uploading metadata to OSF
Hi @jcolomb,

As part of [joss-review](https://github.com/openjournals/joss-reviews/issues/2282) I was wondering if the upload of metadata to osf needs to be made clearer. For instance the osf [documentation](https://docs.ropensci.org/osfr/articles/auth) states that an OSF PAT is required. You are already using one but I needed to look through the code to find this? Also wondering what @mikerspencer thinks about this?
