### 2020-07-02 #2282 @jcolomb
no, this is a "in-house" categorisation that was (so far) not published.
