### 2020-07-02 jcolomb/HCS_analysis#20 @jcolomb
good point, especially since I just had to change the code to be able to make the upload working with the new osfr package.
