### 2020-07-02 067f70c
add info on osf publication
