### 2020-07-02 jcolomb/HCS_analysis#20 @jcolomb
I added some information in the readme file, is it the good place or would you better see it also somewhere else?
