### 2020-07-02 jcolomb/HCS_analysis#20 @aj2duncan
I like what you've put in the Readme but wonder if some information should also go in `Metadata_information/Readme.md` too. I was also wondering if some screen shots might help users too - I found the main app quite intuitive to use, the metadata one slightly less so - might just be me though.
