### 2020-07-02 #2282 @aj2duncan
> no, this is a "in-house" categorisation that was (so far) not published.

Thanks for the clarification.
