### 2020-07-02 jcolomb/HCS_analysis#20 @jcolomb
The metadata one is indeed a bit more complex, and I spent less time on it, too. Thanks for the feedback.
