### 2020-07-08 #2282 @aj2duncan
@danielskatz everything seems to be moving along ok for me.
