### 2020-07-14 jcolomb/HCS_analysis#18 @aj2duncan
Hi @jcolomb, just wondering if there had been any progress on this yet? Thanks.
