### 2020-07-14 jcolomb/HCS_analysis#20 @aj2duncan
Hi @jcolomb, just wondering if you had decided to add any other additional changes to the documentation? Thanks.
