### 2020-07-15 jcolomb/HCS_analysis#20 @jcolomb
yeah, sorry, got this down my todo list.. I still remember having made a video, but cannot find it anymore.

That is quite some work, but I will get to it next week.
