### 2020-07-15 jcolomb/HCS_analysis#18 @jcolomb
still working on that, tried to redo everything from scratch, but it still gives error. Starting to think packrat is just a waste of time grrrr
