+++ paper.md
@@ -90,6 +90,6 @@ Funded by the Deutsche Forschungsgemeinschaft (DFG, German Research Foundation) 
 
 # Dependencies
 
-The software was buildt on R ressources [@R-base]. This work would not have been possible without the  tidyverse environment [@tidyverse; @stingr], packages for interactive processing [@shinyfiles; @shiny; @plotly], statistical analysis [@svm; @rf; @Hmisc; @ica; @glmpath; @R-rstatix; @R-coin] and graphical interface [@RGraphics; @gridExtra; @plotly]. It also depended on the osfr package, which was still in development [@R-osfr] and loaded via the devtools package [@devtools]. We used the packrat package [@Rpackrat]  to dock the project.
+The software was buildt on R ressources [@R-base]. This work would not have been possible without the  tidyverse environment [@tidyverse; @stingr], packages for interactive processing [@shinyfiles; @shiny; @plotly], statistical analysis [@svm; @rf; @Hmisc; @ica; @glmpath; @R-rstatix; @R-coin] and graphical interface [@RGraphics; @gridExtra; @plotly]. It also depended on the osfr package [@R-osfr]. We used the renv package [@R-renv] to dock the project. Other dependencies can be access in the renv.lock file.
 
 # References
