+++ test.bib
@@ -1,187 +1,150 @@
- @Manual{Rpackrat,
-    title = {packrat: A Dependency Management System for Projects and their R Package
-Dependencies},
-    author = {Kevin Ushey and Jonathan McPherson and Joe Cheng and Aron Atkins and JJ Allaire},
-    note = {R package version 0.4.8-56},
-    url = {https://github.com/rstudio/packrat/},
-  }
-
-@Manual{RGraphics,
-  title = {RGraphics: Data and Functions from the Book R Graphics, Second Edition},
-  author = {Paul Murrell},
+@Manual{R-abind,
+  title = {abind: Combine Multidimensional Arrays},
+  author = {Tony Plate and Richard Heiberger},
   year = {2016},
-  note = {R package version 2.0-14},
-  url = {https://CRAN.R-project.org/package=RGraphics},
+  note = {R package version 1.4-5},
+  url = {https://CRAN.R-project.org/package=abind},
 }
-@Manual{gridExtra,
-  title = {gridExtra: Miscellaneous Functions for "Grid" Graphics},
-  author = {Baptiste Auguie},
-  year = {2017},
-  note = {R package version 2.3},
-  url = {https://CRAN.R-project.org/package=gridExtra},
-}
-
-@Manual{stingr,
-  title = {stringr: Simple, Consistent Wrappers for Common String Operations},
-  author = {Hadley Wickham},
-  year = {2017},
-  note = {R package version 1.2.0},
-  url = {https://CRAN.R-project.org/package=stringr},
-}
-
-
-@Manual{tidyverse,
-  title = {tidyverse: Easily Install and Load the 'Tidyverse'},
-  author = {Hadley Wickham},
-  year = {2017},
-  note = {R package version 1.2.1},
-  url = {https://CRAN.R-project.org/package=tidyverse},
-}
-
-@Manual{R-osfr,
-  title = {osfr: API to the Open Science Framework},
-  author = {Chris Hartgerink and VP Nagraj and Ryan Hafen and Julien Colomb},
-  note = {R package version 0.1.0},
-  year = {2017},
-}
-  
-@Manual{Hmisc,
-  title = {Hmisc: Harrell Miscellaneous},
-  author = {Frank E {Harrell Jr} and with contributions from Charles Dupont and many others.},
-  year = {2017},
-  note = {R package version 4.1-0},
-  url = {https://CRAN.R-project.org/package=Hmisc},
-}
-
-@Manual{survival-package,
-  title = {A Package for Survival Analysis in S},
-  author = {Terry M Therneau},
-  year = {2015},
-  note = {version 2.38},
-  url = {https://CRAN.R-project.org/package=survival},
-}
-
-@Manual{svm,
-  title = {e1071: Misc Functions of the Department of Statistics, Probability
-Theory Group (Formerly: E1071), TU Wien},
-  author = {David Meyer and Evgenia Dimitriadou and Kurt Hornik and Andreas Weingessel and Friedrich Leisch},
-  year = {2017},
-  note = {R package version 1.6-8},
-  url = {https://CRAN.R-project.org/package=e1071},
-}
-@Manual{ica,
-  title = {ica: Independent Component Analysis},
-  author = {Nathaniel E. Helwig},
-  year = {2015},
-  note = {R package version 1.0-1},
-  url = {https://CRAN.R-project.org/package=ica},
-}
-@Article{rf,
-  title = {Classification and Regression by randomForest},
-  author = {Andy Liaw and Matthew Wiener},
-  journal = {R News},
-  year = {2002},
-  volume = {2},
-  number = {3},
-  pages = {18-22},
-  url = {http://CRAN.R-project.org/doc/Rnews/},
-}
-
[…]
