+++ test.bib
@@ -696,14 +696,6 @@ OpenSSL},
   note = {R package version 0.2.8},
   url = {https://CRAN.R-project.org/package=osfr},
 }
-@Manual{R-packrat,
-  title = {packrat: A Dependency Management System for Projects and their R Package
-Dependencies},
-  author = {Kevin Ushey and Jonathan McPherson and Joe Cheng and Aron Atkins and JJ Allaire},
-  year = {2018},
-  note = {R package version 0.5.0},
-  url = {https://CRAN.R-project.org/package=packrat},
-}
 @Manual{R-pander,
   title = {pander: An R 'Pandoc' Writer},
   author = {Gergely Daróczi and Roman Tsegelskyi},
