### 2020-07-15 jcolomb/HCS_analysis#18 @jcolomb
renv seems more stable, will still make some test to #21
