### 2020-07-15 jcolomb/HCS_analysis#18 @jcolomb
It is not running if using R4.0, can you confirm it works with r3.6?
