### 2020-07-17 #2282 @mikerspencer
@danielskatz can I check the scope of the review? How involved should we be as reviewers? I think testing should be done prior to review, but realise that's a grey area on problems picked up in the review.
