### 2020-07-17 #2282 @danielskatz
@mikerspencer  - I don't think I understand what you are asking. You should do as much as possible to verify the criteria in the checklist, subject to reasonable effort, and then report what problems you come across so that the submitter can respond.  I'm confused when you say "I think testing should be done prior to review"  There should be some tests provided by the submitter (either manual or via some script or something else), and part of your job is to **verify** that they work and that you think they are reasonable, given the software, and the [JOSS documentation about testing](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)

Let me know if this helps, and if not, please ask me again.
