### 2020-07-20 jcolomb/HCS_analysis#18 @aj2duncan
I seem to have `renv::restore()` success with R.3.6.3 but with some warnings.

```
Warning messages:
1: In read.dcf(file, ...) :
  Line starting 'randomForest, ...' is malformed!
2: In read.dcf(file, ...) :
  Line starting 'randomForest, ...' is malformed!
3: In read.dcf(file, ...) :
  Line starting 'randomForest, ...' is malformed!
```
