### 2020-07-20 #2282 @mikerspencer
Thanks @danielskatz your comment on verification is helpful. I think working on this issue (https://github.com/jcolomb/HCS_analysis/issues/18) is becoming more like co-creation. Keen to have your experienced view!
