### 2020-07-20 #2282 @danielskatz
Well, ideally everything should work for the reviewer when they first try it, but if that was guaranteed, we wouldn't need any reviews :)

So, yes, sometimes the reviewers will work with the author to test things and this can be somewhat iterative.  If it's too much work, you can ask the author to go off and test with colleagues and come back when they are fairly sure the tests will pass for the reviewers, but sometimes, there is something tricky that does need to be worked out between reviewers and authors too.

Thanks for your help in doing this, and getting this far.
