### 2020-07-20 jcolomb/HCS_analysis#18 @jcolomb
thank you for testing, happy to see this working.
(The warning comes from the "document" file, which may be a remnant of packrat.)

can we set this closed by #21 ?
