### 2020-07-24 jcolomb/HCS_analysis#18 @aj2duncan
I'm happy to close this but I think we might need to get an opinion from @mikerspencer
