### 2020-07-27 jcolomb/HCS_analysis#18 @mikerspencer
My plan is to finish the review this week - I've a little more time than the past few months. Fingers crossed!
