### 2020-07-29 #2282 @danielskatz
👋 @aj2duncan & @mikerspencer - As we're now almost 2 months into the review process, can you each please tell me:
1. Are you waiting on @jcolomb for anything?
2. If so, for what?
3. if not, when do you expect your review to be complete?
