### 2020-08-04 jcolomb/HCS_analysis#22 @aj2duncan
Documentation
Hi @jcolomb,

As part of openjournals/joss-reviews#2282 I think that you need some additional documentation on the functionality of the shiny apps. This is linked with #20. The info in the paper is a little clearer than in the README but I'm still finding it a little difficult to navigate my way around.

The screen shot and visual abstract in the paper could be included in the documentation and as with #20 I think that there needs to be a clear process for a new user to follow. I guess the process will be dependent on whether they need to upload metadata to osf first.

Thanks.
