### 2020-08-04 #2282 @danielskatz
> @danielskatz question for you.
>
> I think a statement of need is missing from the github repo but is present in the paper and @jcolomb has linked the paper from the first section of the README. Could you please comment on whether anything additional is required?
>
> Thanks.

It would probably be good to include some of this statement in the README directly.
