### 2020-08-05 jcolomb/HCS_analysis#22 @jcolomb
Thank you, I hope to get some time to start this today.
