### 2020-08-28 358192d
modify helper function
