### 2020-09-01 3fd60da
modif readme +  helper function
