+++ README.md
@@ -7,20 +7,20 @@ This repository contains software developed to analyse behaviour sequence data o
 
 This repository was developed using Rstudio and has many dependencies. 
 
-# installation
+# Installation
 
-clone or download this repository. In your R console type `packrat::restore()`, it should dowload the necessary dependencies (you may need several tries). If you cannot make this to work, you may try to run analysis/Installation_Rpackages.R.
+Clone or download this repository. Run `source(analysis/Installation_Rpackages.R)` to install dependencies.
 
-Now open the shiny apps in Rstudio and click the run button or type `shiny::runApp('analysis/shiny__Analyse_data')`.
+Now open the shiny apps in Rstudio and click the run button or type `shiny::runApp('analysis/shiny__Analyse_data')`. You will open the data analysis software to analyse predetermined datasets.
 
-you may need to first install the command line tools on osx, run this in the terminal
+You may need to first install the command line tools on osx, run this in the terminal
    ` xcode-select --install  `
 
 # Generalities
 
-Two Shiny apps are present in this repository. You will find them in the analysis folder. One does read and check your metadata before pushing the project metadata on osf, the second one is analysing data (master project metadata have no be made public first). One can also use the `master_noshiny.r` script to run the analysis.
+Two Shiny apps are present in this repository. You will find them in the analysis folder. One does read and check your metadata before pushing the project metadata on osf, the second one is analysing data (master project metadata have no be made public first). One can also use the `master_noshiny.r` script to run the analysis, but one has to set varaiables manually.
 
-The repository also contains data to test new functions, information about how to add data and metadata from you project. Dependencies have been taken cared of using the packrat package. Additional information and code which were used to write a paper (not published yet) are also present.
+The repository also contains data to test new functions, information about how to add data and metadata from you project. Dependencies have been taken cared of using the renv package. Additional information and code which were used to write a paper (not published yet) are also present.
 
 
 # Data analysis overview
@@ -35,13 +35,15 @@ The software then performs multidimensional analyses on this latter data to plot
 
 - a project metadata: each project is listed in one spreadseets (saved on osf, see below), path toward other metadata file is indicated there.
 - a lab metadata: here is stocked information about the room where the experiment was performed (daily light cycle is indicated there).
-- an experiment metadata: each row represent one test. Information about the animal tested is stored there.
+- an experiment metadata: each row represent one test session. Information about the animal tested is stored there.
 
 We are using the osf to create a "repository" for the project metadata information (the master metadata file is hosted there because we can read and update it from R). The shiny app was given access to a particular repostiory using [Token Auth](https://cran.r-project.org/web/packages/osfr/vignettes/auth.html) in the osf APIv2 (via the osfr package).
 
 # Using your data
 
-Please refer to https://github.com/jcolomb/HCS_analysis/blob/master/Metadata_information/Readme.md to create new metadata and modify your data to obtain a usable state. Once your data is in a format accepted by the software, you might upload it on github and archive it on zenodo, or keep it locally. By indicating the folder as the raw data folder in github, the software can access the data directly on Github. Once the project metadata has been pushed to osf via the app, you will be able to use the analysis application.
+Please refer to https://github.com/jcolomb/HCS_analysis/blob/master/Metadata_information/Readme.md to create new metadata and modify your data to obtain a usable state. **You will gain a lot of time if you do this before data acquisition**. Once your data is in a format accepted by the software, you might upload it online, or keep it locally. By indicating the folder as the raw data folder in github or gitlab, the software can access the data directly on Github. Once the project metadata has been pushed to osf via the app, you will be able to use the analysis application.
+
+Please make sure you keep information about the time at which the video has been started (it should be indicated in the video name if you are using the HomeCageScan software.)
 
 
 
