### 2020-09-02 9f9a0f0
debug svm with 3 groups and different size
