### 2020-09-02 75df5c1
change default behavior to no svm
