### 2020-09-02 3031827
working with 3 and more than 3 groups
