### 2020-09-02 jcolomb/HCS_analysis#20 @jcolomb
Doing this lead to the discovery of some bugs. I hope they are all taken care of in #23.
Will now finish to work on the documentation.

In addition, working with a different dataset and trying to push it online, made me discover other problems in the analysis application itself. I will run extra test on #24
