### 2020-09-02 #2282 @jcolomb
hopefully done with the difficult parts, hope to finish tomorrow.
