### 2020-09-03 090456b
adding figures and modify part of the text in readme
