+++ Metadata_information/Readme.md
@@ -7,20 +7,26 @@ So far (January 2020), we tested data from different labs, quality and formats. 
 
 Get a version of this software on your desktop! You may install R and Rstudio at this stage. Fork and clone the project if you like to have your data on this github repository, you may just download it otherwise.
 
-Refer to the "Metadata_information/explanation_of_metadata.csv" file for a description of the metadata fields in the different metadata files.
+Refer to the [Metadata_information/explanation_of_metadata.csv](explanation_of_metadata.csv) file for a description of the metadata fields in the different metadata files.
 
 #  Project_metadata.csv file
 
 - Open the local Project_metadata.csv file (you will find it in the data folder) in excel (or similar sofware like libreoffice).
+![](figures/openPMfile.png)
 - Select all cells and change its format to "text" to avoid any intervention of the program in what you are typing
+![](figures/cellformat.png)
 - take the next number available for the experiment ID
+![](figures/exper_ID.png)
 - Keep information about metadata location empty for now, and keep the file open until further notice.
 - in the source_data column, write "this_github" for now.
+![](figures/thisgithb.png)
 - Enter data about the experiment as requested in each cell. 
 - save the file as a .csv document (separator = ",". In excel click "save as" and use the ".csv(macintosh)" option)
 
 
 # Add your data
+
+![](figures/metadatafiles.png)
 
 - copy the "Template_folder_newexp" folder you will find in the metadata_information folder.
 - rename it and enter that information in the Project_metadata.csv file under "folder_path"
@@ -45,11 +51,19 @@ Refer to the "Metadata_information/explanation_of_metadata.csv" file for a descr
 - A second helper (HELPER_join_animal-info) can combine the information with animal information if you have it in a different spreadsheet. You may also try to sort the spreadsheet in a way to be able to copy-paste information. In the end, you should be able to copy-paste information from the eachfile2.csv files into the animal_metadata file
 - Enter missing information in the animal_metadata file.
 
-Note:
+|add metadata figure|
+|:---|
+| ![](figures/add_metadata_1.png) |
+| Caption: The spreadsheet was created with the helper function, each row correspond to one .mbr file found. Animal ID and date will be entered by hand, as the information is in paper form. The information about animal and experience will be copy-pasted later. **This is the most time consuming and error proned step, be patient.** |
+|---|
+
+
+
+Notes:
 - If animals were tested more than once, add a distinctive information in a new column (treatment for instance).
-- Be patient, this step might take some time, but you will get something clean at the end.
 - The lab_ID column in this file should correspond to the lab_ID information entered in the lab metadata information. 
 - Check the file for inconsistancy and other possible problems.
+
 
 
 # Check
@@ -60,11 +74,18 @@ Note:
 - indicate in which folder the data is, if it is neither online nor in the data folder of the software (if you followed this guide step by step, you can ignore this step).
 - choose your project in the selection box.
 - push the "test metadata" button.
-- if they are error messages, investigate and modify the data or metadata accordingly, repeat previous steps
+- if they are error messages, investigate and modify the data or metadata accordingly, repeat previous steps.
 - if they are no error, push the button: "if everything is correct, create the min file", follow the console in Rstudio to make sure it is successful.
 - **Do not click the push button yet.**
 
-# Push
+
+|Check and upload|
+|:---|
+| ![](figures/checker.png) |
+| Caption: The application test that the metadata is in a good format (for instance dates are in the ISO yyyy-mm-dd format), and that every file indicated exists. It also show files existing, but not referenced in the metadata, check at that moment whether some animals were tested and not incorporated in the metadata. The list is also opened in a csv file on Rstudio.  |
+|---|
+
+# Push the data to osf
 
 
 - Your data is ready, but it needs to find a home. You can either keep it offline or bring it online.
@@ -75,21 +96,27 @@ Note:
[…]
