### 2020-09-03 jcolomb/HCS_analysis#20 @jcolomb
@aj2duncan #27 is ready for review, many changes were done to the documentation and to the app.
