### 2020-09-03 #2282 @jcolomb
I am basically done, lots of debug and extension for online data in mbr format, documentation in readme and in metadata_information/readme were reworked quite a bit.
Changes are in the https://github.com/jcolomb/HCS_analysis/tree/joss_reviewanswers branch, and comment may be done in the pull request https://github.com/jcolomb/HCS_analysis/pull/27

Thank you a lot for all the feedback.
