### 2020-09-03 jcolomb/HCS_analysis#20 @aj2duncan
Thanks @jcolomb, I will try to have a look as soon as I can.
