### 2020-09-08 #2282 @danielskatz
> Cheers!

Does that translate into American as "I'll get back to my review soon"? 🙂
