### 2020-09-13 jcolomb/HCS_analysis#20 @aj2duncan
Hi @jcolomb,

I think that the documentation, particularly on the metadata is far more comprehensive and will undoubtedly make it easier for users.

However, I'm getting a few additional errors. First on the test and upload app using the Ro_testdata and pressing the *Test_metadata* button produces the following error:

```
[1] "some files are not grabbed by the metadata, are some animals missing?"
Joining, by = "filepath"
Joining, by = "filepath"
Warning: Unknown or uninitialised column: `sourcce_data`.
Warning: Error in if: argument is of length zero
```

In the analyse data app, choosing the time windows and the multidimensional analysis seems to work well. However, there seems to be a problem with the summary reports. First, there is an error produced on start up. I think that this will be a shiny artefact and not actually have any effect on performance.

![image](https://user-images.githubusercontent.com/10415752/93027303-ea08c700-f603-11ea-99ed-11d9bc91536e.png)

and trying to produce the hourly summary data plots produces this error

```
  Warning: Error in pdf: cannot open file '../data/Ro_testdata/BSeq_analyserfatal: not a git repository (or any of the parent directories): .git/hourly_plotting_corrected.pdf'
```

Note that I downloaded the directory as a zip file rather than cloning it.
