### 2020-09-14 jcolomb/HCS_analysis#20 @jcolomb
Test_metadata button error: I cannot reproduce the problem, I also cannot find "sourcce_data" text anywhere in my downloaded version. Maybe there was a problem during download or you changed something inadvertently after the download (it seems an extra "c" was introduced somewhere...) ?

The typo was indeed there in my version too, I have no clue why it worked on my side. thanks
