### 2020-09-14 9a2b766
Add something for version if the app is downloaded (no access to git release number)
