### 2020-09-14 a7492a4
add default plot to hour summary tab
