### 2020-09-14 jcolomb/HCS_analysis#20 @jcolomb
I added a default plot to avoid the Erro: no applicable method... at startup.
I added a default version number for cases the git repository is downloaded and not cloned, this should allow the software to save things at a right place, I will now check whether the error is gone.
