### 2020-09-14 f3000e2
work version number, copied tags on non git part
