### 2020-09-14 3153d73
modify to allow answer in interactive session
