### 2020-09-14 jcolomb/HCS_analysis#20 @jcolomb
After some cleaning of data left over by older version of the data/software, it seems that everything is now working fine.
(need to use the https://github.com/jcolomb/HCS_analysis/tree/joss_reviewanswers version)
