### 2020-09-14 jcolomb/HCS_analysis#20 @aj2duncan
Hi @jcolomb, thanks for the fixes.

The analyse app is working fine. Still having the same issue with the metadata app but I've also found the problem (and I apologise as I should have noticed the line number reported before). There is a typo on line 392 of the shiny app file. This is where `sourcce_data` was coming from. Fixing the typo allows it to work and everything seems to be working at least for some of the datasets.

Choosing the Rosenmund dataset got me the following error

```
Warning: Error in : 'NA/rosenmund_2015/metadata/rosenmund2015_meta_all.csv' does not exist in current working directory
```

Do you want to allow all the datasets to be testable by the metadata app?
