### 2020-09-14 jcolomb/HCS_analysis#22 @aj2duncan
As far as I'm concerned this can be closed as the documentation is much improved.
