### 2020-09-14 #2282 @aj2duncan
Hi @danielskatz, just to let you know that although there are a couple of small things to be ironed out I'll be done once we close jcolomb/HCS_analysis#20
