### 2020-09-15 jcolomb/HCS_analysis#20 @jcolomb
It seems you tried to look at a dataset that you  have not provided. I added a message in the analysis app for dataset one has to provide locally.

I added a small message in the metadata app too.
