+++ analysis/shiny__testandupload_data/app.R
@@ -389,7 +389,7 @@ server <- function(input, output, session) {
     
       Outputs = paste(WD,Projects_metadata$Folder_path,"Routputs", sep="/")
       #for online projects, outputs are written on disk:
-      if (Projects_metadata$sourcce_data == "https:/") Outputs = paste("../Routputs",Projects_metadata$Folder_path, sep="/")
+      if (Projects_metadata$source_data == "https:/") Outputs = paste("../Routputs",Projects_metadata$Folder_path, sep="/")
 
       values$dir = Outputs
   })
