### 2020-09-15 78af66d
added message for data with source is USB_stick
