### 2020-09-17 jcolomb/HCS_analysis#20 @aj2duncan
Hi @jcolomb,

Thanks for the additional note (btw small typo in column). I think your walkthrough of everything on the metadata README is very clear but I wonder if you might want to consider adding an additional message for the user.

From my reading of things the dropdown for projects to analyse is produced from the `Project_metadata.csv` file which are the entries from the osf. Some of these datasets are also stored (at least partially) locally - R0_testdata, Steele07_HD and Ro_testdata_mbr. So a user can click *test_metadata* on one of those datasets and the app will run but if the dataset doesn't exist (any of the others), it will quit with an error. I would suggest that the user should get some sort of message - rather than the app quitting.

However, the app does function as you intended so I'm happy to close this issue.
