### 2020-09-17 ceed833
added message, cleaning some remnant comments in the app code
