### 2020-10-13 jcolomb/HCS_analysis#18 @mikerspencer
Does packrat::restore run with R 4 yet? I'm trying today but getting an error between dplyr and plogr.
