### 2020-10-13 jcolomb/HCS_analysis#18 @jcolomb
I could not make packrat to work and move the project to the novel renv package environment.
I did not try to make it work with R4.0, so the installation code just install the packages for R4.0 and use renv for R3.6.

This is in the https://github.com/jcolomb/HCS_analysis/tree/joss_reviewanswers branch.
