### 2020-10-14 jcolomb/HCS_analysis#30 @mikerspencer
Inconsistency in file names
There is a general inconsistency in file names (some end `.r` some `.R`), which is causing problems for file loading. e.g. https://github.com/jcolomb/HCS_analysis/issues/28.

Another example: line 447 in `shiny__Analyse_data/app.R` is `source ("Rcode/get_behav_gp.r")`, but the file is `Rcode/get_behav_gp.R`.

Consistent file names would help avoid these mistakes.
