### 2020-10-14 jcolomb/HCS_analysis#18 @mikerspencer
I'm pausing my review while the file name errors are fixed (https://github.com/jcolomb/HCS_analysis/issues/30). Not currently possible to run the app while they exist.
