### 2020-10-15 #2282 @mikerspencer
I've paused my review while I wait for this issue to be fixed: https://github.com/jcolomb/HCS_analysis/issues/30.
