### 2020-10-19 a8ae83b
Delete PCA_strategy.R
