### 2020-10-19 e1eb6de
Delete Timewindows_utilisation.R
