### 2020-10-19 e780580
Delete checkmetadata.R
