### 2020-10-19 71d10bb
Delete create_minfile.R
