### 2020-10-19 09e78f1
Delete grouping_variables.R
