### 2020-10-19 818d248
Delete multidimensional_analysis_svm.R
