+++ analysis/Rcode/multidimensional_analysis_svm.R
@@ -1,128 +1,0 @@
-# #-------------------------SVM analysis
-
-##---testset validation if enough data
-if (testvalidation){
-  #---- Determine the trainset and testset of data, more complex if there is a confounding factor.
-  
-  if (!is.na(Projects_metadata$confound_by)) {
-    Input = Multi_datainput_m2 #take the data with confounding variable
-    
-    L = levels(Input$groupingvar)
-    L2 = levels(Input$confoundvar)
-    Glass = Input %>% filter (groupingvar == L[1])
-    Glass2 = Input %>% filter (groupingvar == L[2])
-    # message if group of different size
-    if (nrow (Glass) != nrow (Glass2))
-      print("the groups do not have the same size !")
-    
-    # split each randomly over confoundvar
-    
-    Glass_1 = Glass %>% filter (confoundvar == L2[1])
-    Glass_2 = Glass %>% filter (confoundvar == L2[2])
-    Glass2_1 = Glass2 %>% filter (confoundvar == L2[1])
-    Glass2_2 = Glass2 %>% filter (confoundvar == L2[2])
-    
-    # prepare datasets
-    indexl = min(nrow(Glass), nrow(Glass2))
-    index     <- 1:indexl
-    trainlength = max(4,  trunc(length(index) / 3)) #max fct deprecated, this run only if there is enough data
-    indexl2 = min(nrow(Glass_1),
-                  nrow(Glass_2),
-                  nrow(Glass2_1),
-                  nrow(Glass2_2))
-    index     <- 1:indexl2
-    trainindex <- sample(index, trainlength)
-    
-    
-    trainset  <-
-      rbind(Glass_1[trainindex, ], Glass_2[trainindex, ], Glass2_1[trainindex, ], Glass2_2[trainindex, ])
-    testset <-
-      rbind(Glass_1[-trainindex, ], Glass_2[-trainindex, ], Glass2_1[-trainindex, ], Glass2_2[-trainindex, ])
-    trainset  <- trainset %>% select (-confoundvar)
-    testset <- testset %>% select (-confoundvar)
-    
-  } else { # if no counfounding variable, it is faster!
-    Input = Multi_datainput_m
-    L = levels(Input$groupingvar)
-    Glass = Input %>% filter (groupingvar == L[1])
-    Glass2 = Input %>% filter (groupingvar == L[2])
-    if (nrow (Glass) != nrow (Glass2))
-      print("the groups do not have the same size !")
-    indexl = min(nrow(Glass), nrow(Glass2))
-    
-    # split each randomly
-    index     <- 1:indexl
-    trainindex <- sample(index, max(8, 2 * trunc(length(index) / 3)))
-    trainset  <- rbind(Glass[trainindex,], Glass2[trainindex,])
-    testset <- rbind(Glass[-trainindex,], Glass2[-trainindex,])
-  }
-  ## we now have trainset and  testset!!
- ######--------------------------------run SVM analysis
-  
-  ##---- getting rid of variables with no variability (all 0)
-  temp = trainset %>% select (-groupingvar)
-  cfreq <- colSums(temp)
-  E1 = names(temp[, cfreq == 0])
-  temp = testset %>% select (-groupingvar)
-  cfreq <- colSums(temp)
-  E2 = names(temp[, cfreq == 0])
-  trainset = trainset [,!names(trainset) %in% c(E1, E2)]
-  testset = testset [,!names(testset) %in% c(E1, E2)]
-  Input = Input[,!names(Input) %in% c(E1, E2)]
-  
[…]
