### 2020-10-19 97584a5
Delete inputdata.R
