### 2020-10-19 4684dac
cleaning .R files on web (#31)
* cleaning .R files on web
