### 2020-10-19 0ceec99
Update and rename app.R to app.r
also got rid of extra call to softwareheader.r
