### 2020-10-19 7478cb4
Rename Softwareheader.R to Softwareheader.r (#32)
rename apps and call to .R to .r codes
