+++ analysis/shiny__testandupload_data/app.r
@@ -1,6 +1,4 @@
-#ANALYSING DATA
-
-
+#Test and upload metadata
 
 # This is a Shiny web application. You can run the application by clicking
 # the 'Run App' button above.
@@ -9,24 +7,23 @@
 #
 #    http://shiny.rstudio.com/
 #
+
+library(shiny)
+require (shinyFiles)
 setwd("../")
 
-versions =list.files("../.git/refs/tags")
-if (length(versions) == 0) {versions =list.files("../tags")}
+versions =try(system("git tag", intern = TRUE))
+if (class(versions)== "try-error"){
+  versions =list.files("../.git/refs/tags")
+}
 version = versions[length(versions)]
-library(shiny)
-library(plotly)
-require (shinyFiles)
-library (rstatix) #effect size calculation
-library (coin) #effect size calculation
-#setwd("analysis")
 ##multidimensional analysis:
 library (randomForest)
 library (ica)
 library (e1071) #svm
 require(Hmisc)   #binomial confidence 
 
-#library(osfr) ##access to osf
+library(osfr) ##access to osf
 
 #normal libraries:
 library (tidyverse)
@@ -34,15 +31,15 @@ library (stringr)
 #for plotting
 library(gridExtra)
 library(RGraphics)
-source ("Rcode/functions.r")
-source ("../Softwareheader.R")
 
-source <- function (file, local = TRUE, echo = verbose, print.eval = echo, 
-          exprs, spaced = use_file, verbose = getOption("verbose"), 
-          prompt.echo = getOption("prompt"), max.deparse.length = 150, 
-          width.cutoff = 60L, deparseCtrl = "showAttributes", chdir = FALSE, 
-          encoding = getOption("encoding"), continue.echo = getOption("continue"), 
-          skip.echo = 0, keep.source = getOption("keep.source")) 
+
+
+source= function (file, local = TRUE, echo = verbose, print.eval = echo, 
+                  exprs, spaced = use_file, verbose = getOption("verbose"), 
+                  prompt.echo = getOption("prompt"), max.deparse.length = 150, 
+                  width.cutoff = 60L, deparseCtrl = "showAttributes", chdir = FALSE, 
+                  encoding = getOption("encoding"), continue.echo = getOption("continue"), 
+                  skip.echo = 0, keep.source = getOption("keep.source")) 
 {
   envir <- if (isTRUE(local)) 
     parent.frame()
@@ -270,260 +267,209 @@ source <- function (file, local = TRUE, echo = verbose, print.eval = echo,
   invisible(yy)
 }
 
-# PMeta = "../data/Projects_metadata.csv" ## use for offline testing
-PMeta = paste0("http://www.osf.io/download/","myxcv")
 
-
+source ("Rcode/functions.r")
+PMeta ="../data/Projects_metadata.csv"
 Projects_metadata <- read_csv(PMeta)
-NO_svm = FALSE
-#Timewindows = data.frame ("run it first")
 
 
 
-# Define UI for application that draws a histogram
-ui <- fluidPage(theme = "bootstrapsolar.css",
-   
-   # Application title
-   titlePanel(title=paste0("BSeq_analyser, ",version))
-   ,source ("../Softwareheader.R")
-   ,     
-   # Sidebar with a slider input for number of bins 
-   sidebarLayout(
-      sidebarPanel(
-         sliderInput("Npermutation",
-                     "Number of permutations to perform for the statistics:",
-                     min = 1,
-                     max = 600,
-                     value = 1)
-         , checkboxInput('RECREATEMINFILE', 'recreate the min_file even if one exists', FALSE)
-         , checkboxInput('perf_SVM', 'Perform the svm analysis (takes time, not always working)', FALSE)
[…]
