+++ analysis/shiny__Analyse_data/app.r
@@ -1,4 +1,6 @@
-#Test and upload metadata
+#ANALYSING DATA
+
+
 
 # This is a Shiny web application. You can run the application by clicking
 # the 'Run App' button above.
@@ -7,23 +9,24 @@
 #
 #    http://shiny.rstudio.com/
 #
-
-library(shiny)
-require (shinyFiles)
 setwd("../")
 
-versions =try(system("git tag", intern = TRUE))
-if (class(versions)== "try-error"){
-  versions =list.files("../.git/refs/tags")
-}
+versions =list.files("../.git/refs/tags")
+if (length(versions) == 0) {versions =list.files("../tags")}
 version = versions[length(versions)]
+library(shiny)
+library(plotly)
+require (shinyFiles)
+library (rstatix) #effect size calculation
+library (coin) #effect size calculation
+#setwd("analysis")
 ##multidimensional analysis:
 library (randomForest)
 library (ica)
 library (e1071) #svm
 require(Hmisc)   #binomial confidence 
 
-library(osfr) ##access to osf
+#library(osfr) ##access to osf
 
 #normal libraries:
 library (tidyverse)
@@ -31,15 +34,14 @@ library (stringr)
 #for plotting
 library(gridExtra)
 library(RGraphics)
+source ("Rcode/functions.r")
 
-
-
-source= function (file, local = TRUE, echo = verbose, print.eval = echo, 
-                  exprs, spaced = use_file, verbose = getOption("verbose"), 
-                  prompt.echo = getOption("prompt"), max.deparse.length = 150, 
-                  width.cutoff = 60L, deparseCtrl = "showAttributes", chdir = FALSE, 
-                  encoding = getOption("encoding"), continue.echo = getOption("continue"), 
-                  skip.echo = 0, keep.source = getOption("keep.source")) 
+source <- function (file, local = TRUE, echo = verbose, print.eval = echo, 
+          exprs, spaced = use_file, verbose = getOption("verbose"), 
+          prompt.echo = getOption("prompt"), max.deparse.length = 150, 
+          width.cutoff = 60L, deparseCtrl = "showAttributes", chdir = FALSE, 
+          encoding = getOption("encoding"), continue.echo = getOption("continue"), 
+          skip.echo = 0, keep.source = getOption("keep.source")) 
 {
   envir <- if (isTRUE(local)) 
     parent.frame()
@@ -267,209 +269,260 @@ source= function (file, local = TRUE, echo = verbose, print.eval = echo,
   invisible(yy)
 }
 
+# PMeta = "../data/Projects_metadata.csv" ## use for offline testing
+PMeta = paste0("http://www.osf.io/download/","myxcv")
 
-source ("Rcode/functions.r")
-PMeta ="../data/Projects_metadata.csv"
+
 Projects_metadata <- read_csv(PMeta)
+NO_svm = FALSE
+#Timewindows = data.frame ("run it first")
 
 
 
-
-
-ui <- fluidPage(
-  
-  # Application title
-  titlePanel("Data uploader for Bseq_analyser: Prepare your home cage monitoring data for analysis."),
-  source ("../Softwareheader.R"),
-  # Sidebar 
-  ".
-  This app may quit abruptly in some cases, for instance if it tries to read data that it does not have access to. Check the message in the R console if it happens. ",
-  sidebarLayout(
-    sidebarPanel(
-      checkboxInput('RECREATEMINFILE', 'recreate the min_file even if one exists', FALSE)
-      , shinyUI(bootstrapPage(shinyDirButton('STICK', "Data_directory", 
-                                             "Choose the directory containing all your HCS data (works only while running the app via Rstudio on your computer):")
[…]
