### 2020-10-19 jcolomb/HCS_analysis#18 @jcolomb
I hope everything is fixed now. You may need to clone the repository again, as git is not really good with case sensitivity.
