### 2020-10-29 jcolomb/HCS_analysis#30 @danielskatz
@mikerspencer - can you confirm this is fixed?  If so, you can continue your review in https://github.com/openjournals/joss-reviews/issues/2282 and @jcolomb can mark this as closed
