### 2020-10-29 #2282 @danielskatz
👋 @mikerspencer  - see https://github.com/jcolomb/HCS_analysis/issues/30#issuecomment-718730338
