### 2020-10-30 jcolomb/HCS_analysis#30 @mikerspencer
Yep, app loads now - great work!
