### 2020-10-30 jcolomb/HCS_analysis#18 @mikerspencer
All seems good now - less packages also makes loading quicker.
