### 2020-11-11 #2282 @mikerspencer
Hi @danielskatz - I agree, it's been a long time. The time I had available when I accepted the review has long gone and it's a shame the work wasn't ready when it was submitted.

My worry and opinion is the code base is extensive (1000s lines) with little obvious testing. I can run the apps as described in the docs, but it's not possible to validate the output which makes me uncomfortable signing the review off. The problem of verification is compounded by the software performing many tasks simultaneously (from a users view), which makes it hard to check whether data retrieval/cleaning/management/analysis/visualisation is happening as expected.

These are large and systematic problems to overcome. I'm aware dealing with them will not be a quick process, but I think it's something that needs to happen to pass review. An alternative, if the author is primarily interested in a citable record of the code is to link the Github repo to Zenodo (or Figshare/etc.).
