### 2020-11-11 #2282 @jcolomb
The software already has a doi via Zenodo (since January 2018), my objective to send it here instead of writing a normal paper (as I did for the CeTrAn software years ago https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0042247, and as I had planned to do in 2018 -I actually had a manuscript ready as one can guess from some files in the repo-) was and still is to make sure the code has no technical flaws, is doing what I think it does, and is enough documented to allow contributions from other researchers. I am not a trained RSE and I take it also as a learning experience.

I am very sorry packrat was not working (anymore) when I submitted the software to JOSS and that it took so long to find a working solution. On the other hand, I think it is difficult to know what is "ready" for being peer reviewed, as software peer review is such a young field.

I am not sure to understand what you call "testing", and thought providing a real test dataset would be sufficient. "Verifications" and "testing" were done on the go (especially by having different datasets that had different problems) and there is indeed no extra code for it, or fabricated data. However, using the master_noshiny.r file as a start, one can do the tasks step by step, and the different steps are creating different variables one can access if the whole code is run. That is how the code was created, worked and debugged...

Are there specific things I can provide to overcome the "large and systematic problems" you mentioned and help with the "validation of the output" ?
