### 2020-11-12 #2282 @mikerspencer
This chapter from Wickham's R Packages books is a really good grounding in testing: https://r-pkgs.org/tests.html.

Like you, I usually rely on informal testing when writing code. But this makes it very difficult for others to review our work. For your own purposes using a test dataset is perfect, but for a reviewer we don't know what the output to be. Writing formal tests (see link above) makes expectations explicit.

I completely agree and sympathise about learning and training for software development - academia is very slow to adopt or encourage better practice.
