### 2020-11-12 #2282 @jcolomb
Thanks, it is indeed some eyes opening: "Always write a test when you discover a bug.".

I will need to learn how to use testing in practice and go through the code again. Indeed, probably a long process.
@danielskatz : I foresee again a long pause in the review process, is it ok to pause the review until I am done with this, probably a couple of months, if schools are not closing again ?
