### 2020-11-12 #2282 @danielskatz
Hi @mikerspencer and @jcolomb  - I think there's some confusion here between best practices and what's required for JOSS.  I really appreciate the discussion, which I think will certainly improve the code, but the question here is if it currently meets the minimum required for JOSS publication.

For example, the testing criterion for JOSS is
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Full unit testing and CI is not needed to pass this, though it would be great if that the code had that.  What's needed at a minimum is a description of the manual steps that someone who installs the code could use to verify that it generally was installed successful and can achieve some set of known good results.

This means that the comment from @mikerspencer
> it's not possible to validate the output which makes me uncomfortable signing the review off. The problem of verification is compounded by the software performing many tasks simultaneously (from a users view), which makes it hard to check whether data retrieval/cleaning/management/analysis/visualisation is happening as expected.

is a problem that needs to be resolved.

@jcolomb - maybe you can propose a set of changes that would allow @mikerspencer to say that the software meets the *minimum* for the JOSS criteria that are still unchecked, and then we can see how much work and how many changes that would be.  We could then either pause this while that work happens, and if @mikerspencer is not available then, I will find another reviewer, or withdraw the submission and start again after that.
