### 2020-11-23 a5fd368
adding comment on filtering time windows
