+++ analysis/Rcode/Timewindows_8.r
@@ -39,8 +39,7 @@ T8 = c("lightcondition", "NIGHT", NA, "nighttime")
 
 #create table with numeric values, values being in minutes
 #Timewindows = data.frame(rbind(T1,T2,T3,T4,T5)) #original split
-Timewindows = data.frame(rbind(T1,T2,T3,T35,T4,T5,T6,T7,T8))
-Timewindows = Timewindows[1:7,]
+Timewindows = data.frame(rbind(T1,T2,T3,T35,T4,T5,T6))
 #if (MAXTIME<daylenght) Timewindows = data.frame(rbind(T1,T2,T3,T35,T4))
 colnames (Timewindows) = c("time_reference", "windowstart", "windowend", "windowname")
 Timewindows$time_reference = as.character(Timewindows$time_reference)
@@ -49,5 +48,6 @@ Timewindows =Timewindows %>%
   mutate (windowend = 60*as.numeric (as.character(windowend)))
 Timewindows = data.frame(rbind(Timewindows,T7,T8))
 
+# erase non-existing time windows (example nightlenght < 6, no middle night)
 Timewindows = Timewindows %>% filter ( is.na(windowend) | windowstart < windowend )
 
