+++ analysis/master_shiny.r
@@ -8,6 +8,7 @@ Multi_datainput_m = Multi_datainput_m[, !colSums(is.na(Multi_datainput_m)) >
 
 #multidimensional analysis, Random forest in 2 rounds result will be used for ICA code below
 source ("Rcode/RF_selection_2rounds.r")# returns RF_selec = Input
+source ("Rcode/ICA.r") # return plot called pls
 
 #multidimensional analysis, PCA followed by wilocoxon test on 1st component, and effect size calculation
 source ("Rcode/PCA_strategy.r")
@@ -24,13 +25,13 @@ Validation_type = ifelse(testvalidation, "independent test dataset", "2-out")
 if (TRUE)
   Acc_sampled = c()
 if (length(unique(metadata$groupingvar)) > 3) {NO_svm = TRUE}
+
 if (nrow(metadata) < 20 || NO_svm) { ## sample size too low, no SVM done
   print("not enough data or too many groups for svm")
   Accuracyreal = NA
   NO_svm = TRUE
   #calcul_text =NA
-  
-  source ("Rcode/ICA.r")
+
   rmarkdown::render ("reports/multidim_anal_variable.Rmd")
   #file.copy("reports/results.rdata", paste0(Outputs,"/multidim_analysis_",groupingby,".Rdata"), overwrite=TRUE,
   #          copy.mode = TRUE, copy.date = FALSE)
@@ -39,11 +40,10 @@ if (nrow(metadata) < 20 || NO_svm) { ## sample size too low, no SVM done
   
 } else{
   if (length(unique(metadata$groupingvar)) == 3) { # case of 3 groups
-    source ("Rcode/ICA.r") 
+
     source ("Rcode/morethan2groups.r")
     rmarkdown::render ("reports/multidim_anal_variable2.Rmd", output_file = "multidim_anal_variable.html")
   } else{ # case of 2 groups
-    source ("Rcode/ICA.r") # return plot called pls
     source ("Rcode/multidimensional_analysis_svm.r") # returns Accuracy (text), Accuracyreal = kappa of result of svm prediction on the test data
     # set
     source ("Rcode/multidimensional_analysis_perm_svm.r") # returns Acc_sampled
