### 2021-01-07 #2282 @jcolomb
thanks, happy new year. We are back in hardcore lockdown (i.e. schools are closed) so it might be delayed a bit more, I will give updates weekly. Sorry again for the delays, and thank you for your patience. @mikerspencer thank you again for the tips, I did the first part if you have time to have a look, "only" the SVM analysis has not been tested. You may want to review here: https://github.com/jcolomb/HCS_analysis/pull/35/files#diff-96933a5a509b8888527a6e745aa92c169f5394a525339682fa1cec926937cf5e
