### 2021-01-17 36e2c80
SVM code taken out of master
