### 2021-01-19 c8ea3ce
independence on metadata, svm tests added
