+++ analysis/Rcode/multidimensional_analysis_perm_svm.r
@@ -1,6 +1,6 @@
 ##-- Acc_sampled should be set a priori to this document, since the code may be run multiple times, we do not set it back to nothing here
 ## testvalidation is set in the master_shiny file (n>15 per group ?)
-
+out_sel_ori = out_sel
 
 if (testvalidation){
   ##inputs are trainset, testset, bestk
