### 2021-01-23 863c9cc
Added mention of reviewer code in readme, add review code for SVM, set tuning range back to default. change some filenames
