### 2021-03-19 jcolomb/HCS_analysis#36 @mikerspencer
Quotes missing from install command
Line 17 in readme is `source(analysis/Installation_Rpackages.r)`, but should be `source("analysis/Installation_Rpackages.r")`.

Submitted as part of JOSS review https://github.com/openjournals/joss-reviews/issues/2282
