### 2021-03-19 jcolomb/HCS_analysis#37 @mikerspencer
error from missing reports/empty.html
Running `shiny::runApp('analysis/shiny__Analyse_data')` from line 19 in the readme, generates a warning:

    Warning in file(con, "r") :
      cannot open file 'reports/empty.html': No such file or directory

Submitted as part of JOSS review: https://github.com/openjournals/joss-reviews/issues/2282
