### 2021-03-24 #2282 @danielskatz
👋 @jcolomb - sorry for the delays here.  @mikerspencer and I have chatted briefly offline so that I could better understand his concerns.  Then I asked a couple of other JOSS editors to take a look as well.  Our opinion is that this is not ready to be published at this point.  @karthik has some specific suggestions, which I've asked him to add here, basically reinforcing what @mikerspencer was uncomfortable about
