### 2021-03-24 #2282 @karthik
👋 @jcolomb
Upon inspection of the submissions, it is not really clear what you are trying to accomplish or how someone else will reuse or benefit from this analysis package. The installation is really complicated, dependencies are unclear (for example, one should never directly import tidyverse in a package setting, but call on specific packages and functions as needed. Otherwise you are importing a massive kitchen sink).

Suggestions for refactoring:
- Write this as a regular R package with exported functions so someone can use it in a script of their own
- Include the shiny app inside the package, so it can serve as a demo.
- The background is not clear enough. Please rewrite the readme so it is written for someone who doesn't already work in your lab.
- I think more work is needed (as you say *We hope to make the software flexible enough to analyse data coming for other similar setups.*) before this would be general enough for a JOSS audience.

Best,
Karthik
