### 2021-03-25 #2282 @mikerspencer
Hi @jcolomb,

Some comments to help with subsequent work:

* Have a look at the Wickham & Bryan book on R packages (https://r-pkgs.org/). It's got some really helpful material on structuring work and testing.
* Try and break down work/repos into sections, this will make it easier for others to follow logic and find the parts they need.
* Focus the repo on a task and remove code that no longer contributes to the goal. This will likely evolve over time.

Best of luck with future work,
Mike
